+++ tests/tile_grid/test_tile.py
@@ -1,7 +1,6 @@
 from tests.interface.tile import Tile
 from tests.utils.html import assert_valid_html, tile_has_anchor, tile_has_img
 from tests.utils.filters import mock_markup_filter
-from terminal.plugins.md_to_html.plugin import DEFAULT_MARKUP_FILTER_NAME
 from tests import defaults
 import pytest
 TILE_MACRO_PATH = "pluglets/tile_grid/templates/j2-macros/tile.j2"
@@ -10,12 +9,6 @@ TILE_MACRO_PATH = "pluglets/tile_grid/templates/j2-macros/tile.j2"
 @pytest.fixture
 def tile_macro(env_with_terminal_loader):
     return env_with_terminal_loader.get_template(TILE_MACRO_PATH)
-
-
-@pytest.fixture
-def env_without_markup_filter(env_with_terminal_loader):
-    del env_with_terminal_loader.filters[DEFAULT_MARKUP_FILTER_NAME]
-    return env_with_terminal_loader
 
 
 class TestTile():
@@ -80,19 +73,5 @@ class TestTile():
         rendered_tile = tile_macro.module.make_tile(valid_linked_image_tile)
         expected_figcaption = mock_markup_filter(context={}, value="myCaption")
         expected_html = "<figcaption>" + expected_figcaption + "</figcaption>"
-        assert expected_html in rendered_tile
-        assert_valid_html(rendered_tile)
-
-    def test_no_render_error_if_markup_filter_undefined(self, env_without_markup_filter, valid_linked_image_tile):
-        tile_macro = env_without_markup_filter.get_template(TILE_MACRO_PATH)
-        try:
-            tile_macro.module.make_tile(valid_linked_image_tile)
-        except Exception as ex:
-            pytest.fail(f"Got exception during render: {ex})")
-
-    def test_caption_is_not_run_through_markup_filter_if_md_to_html_plugin_disabled(self, tile_macro, valid_linked_image_tile):
-        valid_linked_image_tile.caption = "myCaption"
-        rendered_tile = tile_macro.module.make_tile(valid_linked_image_tile)
-        expected_html = "<figcaption>myCaption</figcaption>"
         assert expected_html in rendered_tile
         assert_valid_html(rendered_tile)
